Fix test_console_escape in new version of netmiko (#16414)
Description of PR
Summary: Fix test_console_escape in new version of netmiko
Fixes #
The netmiko version was updated in the latest sonic-mgmt container. After that test case tests/dut_console/test_escape_character.py failed as did not match the expected pattern
Latest version:

paramiko: 2.9.5
netmiko: 3.2.0
Root cause

    duthost_console.send_command("\x03",
                                 expect_string=r"{} packets transmitted".format(packet_number), max_loops=300)
In newer versions of netmiko, the send_command method will first try to match the command itself, so it will try to match \x03.
When we send \x03 to the terminal, it displays as ^C in the terminal, so in the output we see the string form: ^C

....

\x03 is a single-byte control character (ASCII value 3)
^C is a combination of two characters (^ and C)
When we read the output from the terminal, we get the display form ^C, not the original control character \x03. They are not matched, so case fail

Solution

    # Send interrupt character directly
    duthost_console.write_channel("\x03")
    # Matching the expected output content
    duthost_console.read_until_pattern(pattern=r"{} packets transmitted".format(packet_number))
Avoid using send_command to send the interrupt character. Instead, use write_channel to send the interrupt signal directly, and then use read_until_pattern to match the expected output

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>