[sonic-mgmt][dualtor-120] Fix flakiness of fdb/test_fdb_mac_learning.py
In case of dualtor active-standby, interface flap doesn't have any
effect on mux status and thus there is no need to wait for mux status
consistency. Updating the if condition to wait for mux consistency only
for "dualtor-aa" which also allows delay (time.sleep(30)) needed to fix
flakiness of this test for dualtor-120.