Add check for max root dispersion on the PTF's NTP daemon (#17629)
* Add check for max root dispersion on the PTF's NTP daemon

When using Chrony as the NTP daemon on the DUT, Chrony will not use NTP
sources that have a root dispersion of more than 3 seconds (configurable
via /etc/chrony/chrony.conf, but we currently don't touch that setting).
Therefore, block here until the root dispersion is less than 3 seconds
so that we don't incorrectly fail the test.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Fix math and fix whitespace

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>