Skip zmq/test_gnmi_zmq.py on all Arista platforms (#17524)
What is the motivation for this PR?
The test is written for smartswitch as it enabled this mode unconditionally: https://github.com/sonic-net/sonic-mgmt/blob/master/tests/zmq/test_gnmi_zmq.py#L54-L58

As of now none of the Arista platforms operate in this mode.

How did you do it?
Add a conditional mark to skip this on Arista platforms.

How did you verify/test it?
Ran this test on Arista-7260CX3

zmq/test_gnmi_zmq.py::test_gnmi_zmq SKIPPED (Test is for smartswitch)