[sonic-mgmt][dualtor-120] Fix flakiness of fdb/test_fdb_mac_learning.py (#17640)
What is the motivation for this PR?
Test is failing sometimes with error Failed: After starting Ethernet0 and populating fdb, corresponding mac address entry not seen in mac table

How did you do it?
In case of dualtor active-standby, interface flap doesn't have any effect on mux status and thus there is no need to wait for mux status consistency. Updating the if condition to wait for mux consistency only for dualtor-aa which also allows delay (time.sleep(30)) needed to fix flakiness of this test for dualtor-120.

How did you verify/test it?
Stressed the test and no failures are seen with the fix.