[subnet decap] Send more packets for verification (#17660)
What is the motivation for this PR?
For the IPinIP packets with dst IP in the VLAN subnet and src IP not in the configured decap term src subnet, they should be forwarded downstream to the server node. The testcase runs arp_responder to facilitate the neighbor setup, but some of those IPinIP packets might be dropped by the kernel in this ARP setup stage.

How did you do it?
Send more packets.

How did you verify/test it?
Pass on dualtor.

Signed-off-by: yenlu <yenlu@cisco.com>