see the following syslog errors when "config reload": (#17469)
Description of PR
Ignore netlink socket readData error logs in syslog that are causing some tests to fail

Fixes # (issue)

This error message was ignored earlier, but only from swss. But the message is seen for teamd as well. Since there is no functional impact due to these logs, these are ignored.

More details abot the message are in sonic-net/sonic-swss#353

The fix for swss was done under #13639

But it was not done for teamd container.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>