Merged PR 10736: Updated ptf ip for ixia testbeds to use new ixia container
Updated the PTF IP configuration for Ixia testbeds to utilize the new Ixia container.

The Ixia chassis (10.3.145.74) was upgraded to OS version 10.xx, while the existing Ixia container was still running version 9.xx, resulting in connectivity issues. To resolve this, a new container was created, and the necessary changes have been implemented to ensure the testbeds use the updated container instead of the previous one.

----
#### AI description  (iteration 1)
#### PR Classification
Configuration update for testbeds.

#### PR Summary
This pull request updates the IP addresses for PTF in ixia testbeds to use a new ixia container.
- Changes in `/ansible/testbed.yaml` to update `ptf_ip` from `10.64.246.150/23` to `10.64.246.188/23`.
- Added new DUT entries `str2-7250-sup-1` and `str2-7804-sup-1` in relevant testbed configurations.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->