chore: add option to specify image-tag for ptf/ixia image (#17696)
Description of PR
Summary:
Fixes # (issue) Add option to specify ptf image, in case we need to support different image version per testbed

Approach
What is the motivation for this PR?
We run into this problem where a few testbed needs to change the versioning of ptf (ixia) image. This is not possible with the current add-topo when integrate with pipeline and need to do manually

By adding this option, we can specify the image version in the testbed.yaml file using ptf_imagetag. For example

For testbed that doesn't specify this option, we will fall back to the default that was set in add_topo.

- conf-name: vms-snappi-sonic-multidut
  group-name: vms6-1
  topo: ptf64
  ptf_image_name: docker-ptf-snappi
  ptf_imagetag: 10.25.2405.106 # Specify the image version here
  ptf: snappi-sonic-ptf
  ptf_ip: 10.251.0.232
  ptf_ipv6:
  server: Server_6
  vm_base:
  dut:
    - sonic-s6100-dut1
    - sonic-s6100-dut2
  inv_name: snappi-sonic
  auto_recover: 'True'
  comment: Batman

Signed-off-by: Austin Pham <austinpham@microsoft.com>