Set mgmt branch to build branch in PR test (#17754)
What is the motivation for this PR?
In current PR test, we are using hard code "master" for MGMT_BRANCH param, it would increase difficulty when we create a new release branch. For example, in current 202411 branch, the MGMT_BRANCH param was set to "master", which is incorrect.

How did you do it?
1.Set MGMT_BRANCH to $(BUILD_BRANCH) in PR test, then request cherry-pick to 202411 branch.
2.Add blank space in SPECIFIC_PARAM to resolve vs code yaml format extension error.
3.Do not use system defined if condition to get MGMT_BRANCH

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>