Add support for Ubuntu 24.04 as the host and sonic-mgmt container (#16039)
Approach
What is the motivation for this PR?
When running with Ubuntu 24.04 host, many tasks from a sonic-mgmt container may fail. For example, running add-topo for a KVM testbed will fail. This is because of a newer version of Python being used in Ubuntu 24.04 that doesn't support some of the packages that ansible modules attempt to import.

How did you do it?
Update the code in sonic-mgmt to add support for Ubuntu 24.04 as a host and for the sonic-mgmt container to be running Ubuntu 24.04.

How did you verify/test it?
Tested on a Ubuntu 24.04 host and sonic-mgmt container based on Ubuntu 24.04.

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>