Update issue link in conditional mark for test_pfcwd_wb (#17806)
What is the motivation for this PR?
test_pfcwd_wb async_storm test was skipped in KVM with issue #16233, the issue was created for socket connection issue, but now it's failing for warmboot-finalizer service timeout, so need to close the issue and create new one.

How did you do it?
Create issue #17803 and update link.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>