Ignore nss_tacplus error message from chrony seen during config reloads (#17801)
PR #16239 ignored an error message coming from ntpd (actually written by
nss_tacplus). As that PR explains, this can happen when a `config reload`
is run, and chrony calls `getpwnam` when `/etc/tacplus_nss.conf` is
still being regenerated. However, it is not a failure.

With the change of NTP daemons, this loganalyzer ignore regex needs to
be updated to match chrony.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>