feat: support trimming lab inv file (#17348)
Description of PR
Support trimming the inventory files such as ansible/lab, ansible/t2_lab etc when passing --trim_inv option.

Summary:
Fixes # (issue) Microsoft ADO 30056122

Approach
What is the motivation for this PR?
When we enable inventory trimming by passing the --trim_inv option, the current logic is to only trim the ansible/veos file, but we noticed that the other inventory file (such as ansible/lab) should also be trimmed because it contains the configs of all the devices in that lab, but we only need the configs related to the current test run. Therefore, we decided to support trimming these inventory files as well.

Please note that the PDU & Fanout hosts trimming is not supported in this PR as it's currently blocked by #17347

How did you do it?
How did you verify/test it?
I ran the new trimming logic on various lab files and can confirm it's working well:

https://elastictest.org/scheduler/testplan/67c7ad505048655bf9cf8a58
https://elastictest.org/scheduler/testplan/67c78be48dcac0cdc64a3998
https://elastictest.org/scheduler/testplan/67c78cc7f60a7a79ff1ae585
https://elastictest.org/scheduler/testplan/67c78c9c8dcac0cdc64a399c
https://elastictest.org/scheduler/testplan/67c7b419d0bae94c81d8a9d6
https://elastictest.org/scheduler/testplan/67ca846a5048655bf9cf8f7b
Any platform specific information?

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>