Add multi-asic support for test-intf-fec (#17814)
Description of PR
Summary: Add multi ASIC support for test-intf-fec. This is possible with the utility command update in sonic-net/sonic-utilities#3819
Fixes # (issue) 28838870

Approach
What is the motivation for this PR?
Described

How did you do it?
Update the command from sonic-net/sonic-utilities#3819 and update the code base so that works with T2. For 202405

Please note that for a release branch to work internally, the following PR here needs to be included:

#17183
#14661
#16424
#15481

How did you verify/test it?
T2 platform verified

Signed-off-by: Austin Pham <austinpham@microsoft.com>

---------

Signed-off-by: Austin Pham <austinpham@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>