[dualtor] update template to latest (#17879)
What is the motivation for this PR?
Old template is not up to date and does not match with changes in vm_topo results. Update it so the generated minigraph work.

How did you do it?
Copy the section from minigraph_dpg.j2

How did you verify/test it?
Run yang validation on generated minigraph.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>