vSonic ipv6 loopback routes need to be /128
cEOS advertises loopback routes as /128, so vSonic should
as well when used as the peer.  This was causing test failures
in things such as snmp/test_snmp_loopback.py when using a
vSonic peer instead of cEOS.  It was instead receiving multiple
/64 routes since each peer has a loopback address using a /128
that is technically in the same /64 range, so there was no
guarantee the packets would be shipped off to the right peer.

Signed-off-by: Brad House <bhouse@nexthop.ai>