[sonic-mgmt][dualtor-aa] Fix flakiness of fdb/test_fdb_mac_learning.py (#17873)
What is the motivation for this PR?
After link bringup, it's taking some time for mux status to be consistent in dualtor-aa topology (i.e SERVER_STATUS is 'unknown'). And it's not a test specific issue, I can see similar behaviour on dut where dualtor-aa is deployed.

How did you do it?
So increasing the timeout to 300 (currently it's 150 secs) to fix flakiness.