Increase timeout to 5 in verify_packet_any_port for background traffic (#17904)
What is the motivation for this PR?
The test is giving us a false negative

msg        = 'Did not receive expected packet on any of ports [7, 13, 17, 30, 27, 25, 5, 34, 21, 16, 24, 1, 33, 12, 4, 20, 2, 0, 11... 01  .............0..
0050  00 AA BB CC DD EE                                ......
==============================
'
self       = <tests.common.plugins.ptfadapter.ptfadapter.PtfTestAdapter testMethod=runTest>

/usr/lib/python3.8/unittest/case.py:753: AssertionError
Although on a closer look we found that the DUTis forwarding the packet in a reasonable duration of time but for some reason testutils.verify_packet_any_port is taking longer to detect it.

There is also another issue which doesn't cause any failure but defeats the purpose of testing. In case of active-active dualtor we call setup_standby_ports_on_rand_unselected_tor_unconditionally to put the system in active-standby mode. If this is called after background_traffic then the background trafffic flows through the unselected ToR which is not desired.

How did you do it?
Increase the timeout to 5s from system default for testutils.verify_packet_any_port

Make the order of fixture execution deterministic so that setup_standby_ports_on_rand_unselected_tor_unconditionally is called before background_traffic

How did you verify/test it?
Verified on Arista-7050CX3 with dualtor-aa topology.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>