Support Ubuntu 24.04 server in KVM (#17883)
What is the motivation for this PR?
When running with Ubuntu 24.04 host, many tasks from a sonic-mgmt container may fail.
One reason is that a newer version of Python being used in Ubuntu 24.04 that doesn't support some of the packages that ansible modules attempt to import. Such as imp module is removed in Python 3.11, need to use importlib, and
PR #16039 did some replacement but still need to replace in somewhere else.
Another reason is that some ansible modules must be running in a python3 venv for scecurity reason.
Also there are some access issues need to fix.

How did you do it?
Try to use import importlib instead of importing imp
Install python3-venv in Ubuntu 24.04 server
Update libvirt qemu configuration in Ubuntu 24.04 server
* fix

* remove

---------

Co-authored-by: xwjiang2021 <96218837+xwjiang2021@users.noreply.github.com>