Increase timeout for gathering sonic hosts facts (#17945)
What is the motivation for this PR?
Gather facts will run commands on sonic hosts and get the results of platform information. But in dualtor, hosts may need more time to get the return results and caused pretest failure.

How did you do it?
Increase timeout in gathering facts and check if the success rate increase in a week.

Signed-off-by: yenlu <yenlu@cisco.com>