Improve arp_responder.py performance (#17280)
* Optimize performance of arp_responder by reusing sockets

Instead of having scapy.sendp create a socket, bind it, send the packet,
and then close the packet every time, just reuse the listening socket
that's been created.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Use libpcap backend for scapy in arp_responder

Use the libpcap backend in arp_responder, which sets up a ring buffer
for incoming packets. This is expected to be more efficient
than making many recvmsg syscalls, especially during ARP/NDP floods.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Enhance BPF filter for arp_responder

Enhance the BPF filter for arp_responder to only get ARP/ICMP packets
for the IPv4/IPv6 addresses that it is interested in. This is especially
useful in large VLANs with many hosts.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>