Changes for VOQ single-dut multi-asic (#17838)
What is the motivation for this PR?
New testbed modification of VOQ Single DUT and multi-asic.
How did you do it?
If switch type is VOQ, decide based on the number of Duts and also based on if the DUT is modular chassis or not.
How did you verify/test it?
Any platform specific information?
Supported testbed topology if it's a new test case?
T2 VOQ Single Dut Multi ASIC

Signed-off-by: yenlu <yenlu@cisco.com>