Add feature check when enable bmp in config_sonic_basedon_testbed.yml  (#18035)
Approach
What is the motivation for this PR?
upgrade_path test pipeline is broken due to this change, since the from version may not support bmp feature.

How did you do it?
Add config cmd check and skip the cli if bmp feature is not supported in the version.

How did you verify/test it?
PR validation will cover the sanity check.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>