[DHCP] Add dhcp_server and dhcp_relay services to critical services if features are enabled (#17976)
What is the motivation for this PR?
Sometimes dhcp containers are messed up by previous test, but sanity check cannot notice it because they are not critical services

How did you do it?
Add them to critical services if features are enabled

How did you verify/test it?
Run tests

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>