Changes to support RDMA Tests for VOQ Chassis (#17763)
What is the motivation for this PR?
In case of a VOQ Chassis, the loss happens on the ingress side instead of the egress side which happens in case of Non- VOQ Chassis.

How did you do it?
By adding a condition for VOQ Chassis.

How did you verify/test it?
Ran it in Microsoft Lab.