[testbed-cli opt] skip snmp_facts when ceos container doesn't exist to save time (#18095)
What is the motivation for this PR?
When adding topology with batch size 30, the task snmp_facts may take hours due to there is no ceos container responding.
We need to skip this task when there is no ceos container to save time.

How did you do it?
Check if ceos container exsit, if no, skip snmp_facts task

How did you verify/test it?
Tested on testbed with add-topo after remove-topo
from the pic, we can see that, due to ceos container was deleted when remove-topo, so when add-topo, there is no ceos container, and the snmp_facts task was skipped.

Signed-off-by: yenlu <yenlu@cisco.com>