Fix Add Cluster Skip condition and add everflow check. (#17972)
Description of PR
Summary:
This PR updates the test_add_cluster logic to ensure it is skipped for platforms that are not multi-ASIC and have a non-Nokia HWSKU. Previously, the test was only skipped for Nokia HWSKUs, which was not correct.

Enhanced verification for applied configurations. This change introduces checks for additional database tables and implements a comparison of the configuration state before and after changes are applied.

Type of change
Add correct condition to skip unverified platform and hwsku.
Enhanced verification

Approach
What is the motivation for this PR?
Skip test_add_cluster when the platform is not multi-ASIC and the HWSKU is not Nokia.
Enhanced verification.
How did you do it?
Fix Skip Condition
Enhanced verification

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>