Increase instance limit to reduce PR checker running time (#18088)
What is the motivation for this PR?
Previously, we capped the maximum number of instances for each PR checker at 25. However, due to increased demand and other factors, this limit has become insufficient, leading to run times exceeding 3 hours. This PR raises the limit to 40 instances to address the issue.

How did you do it?
This PR raises the limit to 40 instances to address the issue.

Signed-off-by: yenlu <yenlu@cisco.com>