[dualtor-io] Running test script on devices to control timing of operations
previous version of PR ran commands sequentially on the device. This did
not accurately simulate the issue because it took time to execute
operations on the device and was not able to accurately simulate the
situation that caused the crash.

This PR instead creates a script dualtor_neighbor_route_share_ip.py,
which runs the operations on both ToRs.

Signed-off-by: Nikola Dancejic <ndancejic@microsoft.com>