add sleep (#18186)
Description of PR
Summary:
Fixes # (issue)
test_acl.py has a high chance to fail on Cisco chassis.
It's a flakiness failure, add extra sleeping time to wait for the routes to be programmed.

Approach
What is the motivation for this PR?
test_acl.py has a high chance to fail on Cisco chassis.
It's a flakiness failure, add extra sleeping time to wait for the routes to be programmed.

How did you do it?
Add extra sleeping time to wait for the routes to be programmed.

How did you verify/test it?
Test on physical testbeds and it stables passes.

Any platform specific information?
Supported testbed topology if it's a new test case?

co-authorized by: jianquanye@microsoft.com

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>