Merged PR 11874: Fix overlapped VLANs introduced by #8293
Fix overlapped VLANs introduced by #8293

PR #8293 was to add graph/inv info for the devices moved to TK5 in 2024.

The PR was created in 2024 and merged in April 2025.
The orignally allocated VLANs were taken by other testbeds
added during this time. After #8293 was merged, it caused
overlaopped VLAN allocations on leaf fanout.

This PR is to fix the problem by allocating new VLANs for
the leaf fanout ports.

After this PR is merged, all the impacted fanout devices
need to be deployed.

The vlan information of the root-leaf fanout ports need to
be updated on the root fanout switch too.

This change also fixed the vlan range for some root-leaf fanout ports.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses the issue of overlapping VLANs introduced in a previous change (#8293) by updating VLAN IDs in the configuration file.
- Updated VLAN IDs in `/ansible/files/sonic_strtk5_links.csv` to resolve overlaps.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->