Fix parsing error and speedup testing in test_srv6_no_sid_blackhole (#18163)
Summary: Previously, the test case verified the dropping of 100 packets one-by-one, which was very slow. Now, we changed it to send 100 packets in a single batch and then verified the no-receiving behavior. Besides, there was a parsing error regarding the counter value when the platform is a mellanox device. I forgot to convert str to int before calculating the counter values.
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>