Skip T2 tests not applicable to disaggregated chassis (#18266)
Approach
What is the motivation for this PR?
Modular chassis specific tests that involve triggers on supervisor are not applicable to disaggregated chassis.

How did you do it?
Add skip conditions for BGP tsa-tsb-services tests that are specific to modular chassis with a supervisor

How did you verify/test it?
Verified that the tests are skipped on disaggregated chassis

Signed-off-by: yenlu <yenlu@cisco.com>