[multi_passwd_ssh] switch ipv6 back to ipv4 when failed (#17127)
When running a command through the plugin multi_passwd_ssh, it will try every password, and at the end, the password will be reset to the original password. If ipv4 fails, the connection error will be caught, and switch to using ipv6, and it will stick to using ipv6 for all future connections. This change will switch back to trying ipv4, if ipv6 fails. Sometimes we have ipv4 and ipv6 recorded in inventory for a device but only one of them (typically ipv4) is configured (like when the mgmt interface is configured manually), the previous behavior is multi_passwd_ssh will always fail if an ipv4 connection ever fails, like during a reboot, or waiting for some breaking network changes.

What is the motivation for this PR?
When ipv4 fails, the multi_passwd_ssh switches to using ipv6. We would want it to continue trying both ipv4 and ipv6, and not just one.

How did you verify/test it?
Run some ansible playbooks

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>