added test documentation for snappi-pfcwd all-to-all testcase (#18229)
Description of PR
Summary:
Added explanation to the PFCWD all-to-all testcase.

@vmittal-msft , @rawal01 , @yatishkoul , @sdszhang for viz.

Fixes #18216

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202012
 202205
 202305
 202311
 202405
 202411
Approach
What is the motivation for this PR?
There was one line description of the test case and needed more explanation about the what the testcase intended to do.

How did you do it?
Added documentation.

How did you verify/test it?
No verification needed as this is added as comments in the test-case description.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>