Remove asic dependency for show queue counters (#18383)
What is the motivation for this PR?
PR #3865 updated the generate_dump script to run the show queue counters command only in the host namespace, instead of for each ASIC namespace.

As a result, the test case in show_techsupport/test_techsupport.py, which was verifying that the command ran for each ASIC namespace failed.

This PR modifies the testcase to align with the new behavior.

Fixes #18364

How did you do it?
Removed {} from {} show queue counters in misc_show_cmds, which is used to generate commands_to_check, to prevent per-ASIC expansion and check only the host namespace.

How did you verify/test it?
run show_techsupport/test_techsupport.py