test_upgrade_path: Fix ceos bgp cmd parse (#18212)
What is the motivation for this PR?
Newer version of eos updates output of show ip bgp neighbor, which changes indentation. This breaks the parsing of restart_time.

How did you do it?
Regex match on integer and ignore indentation.

How did you verify/test it?
Tested on Arista testbed

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>